Refactor Steam controllers and add SteamImageHelper
Refactored SteamMarketController, SteamShopController, and SteamSocialController to instantiate their services internally instead of via constructor injection. Introduced SteamImageHelper for extracting and building Steam item image URLs, and integrated it into SteamMarketService. Improved logger usage to handle null loggers gracefully. Cleaned up and unified service constructors and logging in SteamShopService and SteamSocialService.

TODO: currently, for what ever reason, it stopped loading the .env
